refactor pdf module
