+++ lib/mindee/extraction.rb
@@ -2,4 +2,3 @@
 
 require_relative 'extraction/tax_extractor'
 require_relative 'extraction/multi_receipts_extractor'
-require_relative 'extraction/pdf_extractor'
