--- lib/mindee/extraction/pdf_extractor.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-require_relative 'pdf_extractor/pdf_extractor'
-require_relative 'pdf_extractor/extracted_pdf'
