+++ lib/mindee/pdf.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require_relative 'pdf/extracted_pdf'
 require_relative 'pdf/pdf_compressor'
+require_relative 'pdf/pdf_extractor'
 require_relative 'pdf/pdf_processor'
 require_relative 'pdf/pdf_tools'
