+++ spec/image/image_extractor_spec.rb
