+++ spec/pdf/extracted_pdf_spec.rb
@@ -8,7 +8,7 @@ describe 'Invoice extraction' do
 
   it 'should extract a PDF from an image' do
     jpg_stream = File.open("#{product_data_dir}/invoices/default_sample.jpg", 'r')
-    pdf_wrapper = Mindee::Image::PdfExtractor::ExtractedPdf.new(jpg_stream, 'dummy.pdf')
+    pdf_wrapper = Mindee::PDF::PDFExtractor::ExtractedPDF.new(jpg_stream, 'dummy.pdf')
     expect do
       pdf_wrapper.page_count
     end.to raise_error Mindee::Errors::MindeePDFError
