+++ spec/pdf/pdf_extractor_spec.rb
@@ -20,13 +20,13 @@ describe 'Invoice extraction' do
     jpg_input = Mindee::Input::Source::PathInputSource.new(invoice_default_sample_path)
     expect(jpg_input.pdf?).to eq(false)
 
-    extractor = Mindee::Image::PdfExtractor::PdfExtractor.new(jpg_input)
+    extractor = Mindee::PDF::PDFExtractor::PDFExtractor.new(jpg_input)
     expect(extractor.page_count).to eq(1)
   end
 
   it 'should extract invoices from a PDF (no strict mode)' do
     pdf_input = Mindee::Input::Source::PathInputSource.new(invoice_splitter_5p_path)
-    extractor = Mindee::Image::PdfExtractor::PdfExtractor.new(pdf_input)
+    extractor = Mindee::PDF::PDFExtractor::PDFExtractor.new(pdf_input)
 
     expect(extractor.page_count).to eq(5)
 
@@ -45,7 +45,7 @@ describe 'Invoice extraction' do
 
   it 'should extract invoices from a PDF (strict mode)' do
     pdf_input = Mindee::Input::Source::PathInputSource.new(invoice_splitter_5p_path)
-    extractor = Mindee::Image::PdfExtractor::PdfExtractor.new(pdf_input)
+    extractor = Mindee::PDF::PDFExtractor::PDFExtractor.new(pdf_input)
 
     expect(extractor.page_count).to eq(5)
     expect(loaded_prediction.invoice_page_groups.length).to eq(3)
